+++ ShadowClip/services/IClipCreator.cs
@@ -6,7 +6,7 @@ namespace ShadowClip.services
 {
     public interface IClipCreator
     {
-        Task ClipAndUpload(string originalFile, string clipName, int start, int end,
+        Task ClipAndUpload(string originalFile, string clipName, double start, double end,
             IProgress<EncodeProgress> encodeProgress,
             IProgress<UploadProgress> uploadProgress, CancellationToken cancelToken);
     }
